requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
Werkzeug==2.3.7